chore: update imports and file cleanup
