+++ client/packages/service-message/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./NotificationService"
+export * from "./ServiceMessageIcon"
