--- client/packages/service-message/service-message/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-service-message";
-export * from "../types/types"
